Remove spells-bols.json
We're going to look into re-adding this, as it's something that seemed
interesting at the time. For now, though, I'm going to remove it as some
of the other end-users were bringing it to my attention that the spells
within the Book were either poorly formatted or just didn't work in the
Fifth Edition system. Sorry and thank you for your understanding.