(feat) item data enhancements, Round 1
**data/items.json**
Improved DMG Magic Items: Adamantine Armor - Dwarven Plate.

**data/magicvariants.json**
Added Ammunition of Slaying (DMG calls it Arrow of Slaying but that conflicts with the specific item Arrow of Slaying).

**js/items.js**
Stopped on-the-fly variant creation of basic items that have multiples, e.g. Arrows (20).
Added a Category filter - tried to default it to deselect Specific Variant; it's work-in-progress.
Handled {@lowerName} in-line when generating specific variants that include the name of the specific basic item.
Made SCAG's Studded Armor mundane (it seemed the right thing to do).
DRYed the multiple deselect filter functions into one.
Made references to an item within an item's entries italic per convention.

**js/utils.js**
Updated utils_makeRoller to have die-rollers for " to hit" and "bonus of " entries.