(chore) "use strict" and stricter eslint
Now closer to the eslint "recommended" settings - all our code still passes.

Most JavaScript file changes were to add `"use strict";` as the first line in the file.
Most of the other changes were to add the appropriate `const` or `let` so that all variables were declared.

**Bestiary**
Had to move the imgError function up to the html to work properly with strict; we could disable strict if not happy.
However, the increased checking did pick-up a bug, an orphaned `+"</p>"`, so it was good to find this anyway :).
Postponed the DOM manipulation (appending the <li> content) to further improve initial load performance.

**Editor**
SVG data now in a separate file to avoid a console error.