Fix for slow performance on deleting homebrew sources (#210)
It looks like the issue was that it reserialises the JSON after it
removes each individual item from the homebrew source. For a large source
this can be pretty intensive. I moved the serialisation to outside the two
loops where this happens and it seems to work a lot faster now.

I don't think there's anywhere else that needs to change but I'm
obviously not that familiar with the code so probably worth reviewing
carefully!