Don't _copy hasImages by default (#244)
Require hasImages to be added or copied explicitly.

Most of the time, it doesn't make sense to `_copy` `hasImages` since the
new creature doesn't use the same images. Having to set `hasImages` to
`false` is a bit cumbersome and very easy to forget.

Even if one wants to use the same image as in the source creature, the
fluff entry needs to be created/copied separately anyway and if no fluff
entry is added, the automatically copied `hasImages` results in an empty
images tab.