Creature renderer improvements (#248)
* Add renderer support for prerequisites entries in creature abilities
* Adjust renderer to list creature items above 'top' abilities to match
the usual statblock layout
* Improve creature HP rendering and schema
  * Multiple HP entries (e.g. Hydra) are separated by paragraph breaks
* Rename the `note` property used to specify the body part for multiple
entries to `name` and add the parentheses in the renderer
* Support actual notes for HP entries; listed in parentheses after the
HP value. Currently there is only a single example where this is used
but it is possible that some creatures are missing such notes or have
them placed in the abilities property instead. Paizo themselves are
fairly inconsistent here and mostly separate the HP abilities via `,`
but other times via `;` and occasionally in parentheses like notes...