Remove redundant creature spellcasting name properties (#285)
Delete name when the name can be automatically generated based on type
and tradition instead.

Text converter has been adjusted as well to not generate the name
property in these cases going forward.

NB: We generate the name as `<type> <tradition>` in the renderer which
matches what Paizo usually uses in their statblocks. A small number of
statblocks however invert the order (`<tradition> <type>`). I decided to
also replace those instances with the generated one since that seems
like a fairly insignificant and unintentional difference that is not
worth preserving.

If we want to keep the inverted spelling on the affected statblocks
instead, let me know and I will adjust this PR accordingly.