add Elemental Instinct Barbarian Subclass (#335)
I added the elemental instinct barbarian subclass.

The wording is identical to https://2e.aonprd.com/Instincts.aspx,
however, I slightly changed the sentence containing the link to the
table to make it grammatically correct.

## TODO
- [x] The only remaining issue is that the table containing the
elemental instincts does not link correctly. `npm run serve:dev` shows
the table, however, if I hover over the link, it is not shown as a
popup. Can you help me with that?

---------

Co-authored-by: Constantin Gahr <constantin.gahr@ipp.mpg.de>