Fix/improve `text:` functionality with `@runeItem`s
I was hitting some errors due to `.push()` returning an integer, not the array, so you can't chain `.map()` onto it. However, although this new code returns the correct string, it's still not solving all the search queries I gave it? No idea what's up there but it's too deep for me to care ¯\_(ツ)_/¯